ripd: move leaf-lists to containers of their own
This commit prepares the ground to fix the "passive-interface"
command.

The intention of this structural change in the frr-ripd module
is to make it possible to remove the entire "passive-interface" or
"non-passive-interface" leaf-lists in a single NB_OP_DELETE operation
(by removing the associated container).

This change is in line with what most IETF YANG modules do: enclose
lists and leaf-lists inside containers of their own. This comes
from a guideline from RFC 8407, which says the following:

  "The NETCONF and RESTCONF protocols do not currently support the
   ability to delete all list (or leaf-list) entries at once.  This
   deficiency is sometimes avoided by use of a parent container (i.e.,
   deleting the container also removes all child entries)".

In the future we should probably follow this guideline more closely
by adding new containers for the existing lists and leaf-lists.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>